[frontend] Fixed deletion popup
