+++ opencti-platform/opencti-front/src/private/components/settings/themes/ThemePopover.tsx
@@ -1,16 +1,25 @@
 import { MoreVert } from '@mui/icons-material';
 import { IconButton, Menu, MenuItem } from '@mui/material';
 import React, { FunctionComponent, useState } from 'react';
-import { Disposable } from 'relay-runtime';
+import { Disposable, graphql, RecordSourceSelectorProxy } from 'relay-runtime';
 import { useFormatter } from '../../../../components/i18n';
 import ThemeEdition from './ThemeEdition';
 import { ThemesLine_data$data } from './__generated__/ThemesLine_data.graphql';
-import ThemeDeletion from './ThemeDeletion';
 import Security from '../../../../utils/Security';
 import { KNOWLEDGE_KNGETEXPORT_KNASKEXPORT, KNOWLEDGE_KNUPDATE, KNOWLEDGE_KNUPDATE_KNDELETE } from '../../../../utils/hooks/useGranted';
 import { ThemesLinesSearchQuery$variables } from './__generated__/ThemesLinesSearchQuery.graphql';
 import ThemeType, { deserializeThemeManifest } from './ThemeType';
 import handleExportJson from './ThemeExportHandler';
+import useDeletion from '../../../../utils/hooks/useDeletion';
+import useApiMutation from '../../../../utils/hooks/useApiMutation';
+import { deleteNode } from '../../../../utils/store';
+import DeleteDialog from '../../../../components/DeleteDialog';
+
+const deleteThemeMutation = graphql`
+  mutation ThemePopoverDeletionMutation($id: ID!) {
+    themeDelete(id:$id)
+  }
+`;
 
 interface ThemePopoverProps {
   themeData: ThemesLine_data$data;
@@ -34,6 +43,16 @@ const ThemePopover: FunctionComponent<ThemePopoverProps> = ({
     ...deserializeThemeManifest(themeData.manifest),
   };
 
+  const deleteSuccessMessage = t_i18n('', {
+    id: '... successfully deleted',
+    values: { entity_type: t_i18n('Theme') },
+  });
+  const [commit] = useApiMutation(
+    deleteThemeMutation,
+    undefined,
+    { successMessage: deleteSuccessMessage },
+  );
+
   const handleOpen = (event: React.UIEvent) => {
     setAnchorEl(event.currentTarget);
   };
@@ -45,6 +64,27 @@ const ThemePopover: FunctionComponent<ThemePopoverProps> = ({
   const handleCloseUpdate = () => setDisplayUpdate(false);
   const handleExport = () => {
     handleExportJson(theme);
+  };
+
+  const deletion = useDeletion({ handleClose });
+  const { setDeleting, handleOpenDelete, deleting } = deletion;
+
+  const submitDelete = () => {
+    setDeleting(true);
+    commit({
+      variables: { id: theme.id },
+      updater: (store: RecordSourceSelectorProxy) => deleteNode(
+        store,
+        'Pagination_themes',
+        paginationOptions,
+        theme.id,
+      ),
+      onCompleted: () => {
+        setDeleting(false);
+        handleRefetch();
+      },
+    });
+    handleClose();
   };
 
   return (
@@ -87,19 +127,26 @@ const ThemePopover: FunctionComponent<ThemePopoverProps> = ({
           </MenuItem>
         </Security>
         {!theme.system_default && (
-          <ThemeDeletion
-            id={theme.id}
-            disabled={isCurrentTheme}
-            handleClose={handleClose}
-            handleRefetch={handleRefetch}
-            paginationOptions={paginationOptions}
-          />
+          <Security needs={[KNOWLEDGE_KNUPDATE_KNDELETE]}>
+            <MenuItem
+              onClick={handleOpenDelete}
+              aria-label={t_i18n('Delete')}
+              disabled={isCurrentTheme || deleting}
+            >
+              {t_i18n('Delete')}
+            </MenuItem>
+          </Security>
         )}
       </Menu>
       <ThemeEdition
         theme={theme}
         open={displayUpdate}
         handleClose={handleCloseUpdate}
+      />
+      <DeleteDialog
+        deletion={deletion}
+        submitDelete={submitDelete}
+        message={t_i18n('Do you want to delete this theme?')}
       />
     </div>
   );
